requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-dotenv
sqlalchemy
